requirements.txt
pygame
PyOpenGL
PyGLM
numpy